// generated/src/app/api-models/smrelease-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmreleaseS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  releaseState: number;
  releaseMediatype: number;
  releaseProdfiletype: number;
  releaseLocation: string;
  releaseDescription: string;
  releasePartnumber: string;
  releaseThirdpartyreleasepram: string;
  releaseUserremark: string;
}